Refactor components and reducers to create an Auth module
