+++ react-app/src/js/rootReducer.js
@@ -1,18 +1,11 @@
 import { combineReducers } from "redux";
 
-import registerReducers from "./register-form/reducers";
-import passwordResetReducers from "./password-reset-form/reducers";
-import passwordResetConfirmReducers from "./password-reset-confirm-form/reducers";
-import successMessageReducers from "./welcome/reducers";
-import loginReducers from "./login-form/reducers";
-import { todoReducers, userInfoReducers } from "./dashboard/reducers";
+import authReducers from "./auth/reducers";
+import welcomeReducers from "./welcome/reducers";
+import { todoReducers } from "./dashboard/reducers";
 
 export default combineReducers({
-  register: registerReducers,
-  passwordReset: passwordResetReducers,
-  passwordResetConfirm: passwordResetConfirmReducers,
-  successMessage: successMessageReducers,
-  login: loginReducers,
-  todo: todoReducers,
-  userInfo: userInfoReducers
+  welcome: welcomeReducers,
+  auth: authReducers,
+  todo: todoReducers
 });
